Title: 사각형 채우기 3, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 250417/사각형 채우기 3/rectangle-fill-3.java
@@ -8,7 +8,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
 
-        int[] dp = new int[1001];
+        long[] dp = new int[1001];
         dp[1] = 2;
         dp[2] = 7;
         dp[3] = 22;
